backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
pyScss==1.4.0
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
Werkzeug==3.1.3
